단과대 버튼화면 생성
