+++ AjoudongFE/app/build.gradle
@@ -37,5 +37,6 @@ dependencies {
 
     //noinspection GradleCompatible
     implementation 'com.android.support:design:27.1.1'
+    implementation 'androidx.gridlayout:gridlayout:1.0.0'
 
 }
