+++ AjoudongFE/app/src/main/java/com/example/ajoudongfe/UserMajorSelectActivity.java
@@ -1,53 +1,37 @@
 package com.example.ajoudongfe;
 
+import android.content.Context;
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.MenuItem;
+import android.widget.Toast;
+
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatActivity;
-
 import androidx.appcompat.widget.Toolbar;
-import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.core.view.GravityCompat;
 import androidx.drawerlayout.widget.DrawerLayout;
 
-import android.content.Context;
-import android.content.Intent;
-
-import android.os.Bundle;
-import android.view.MenuItem;
-import android.view.MotionEvent;
-import android.view.View;
-
-import android.widget.ImageButton;
-import android.widget.ImageView;
-import android.widget.Toast;
-
 import com.google.android.material.navigation.NavigationView;
 
-public class UserMainActivity extends AppCompatActivity {
-
+public class UserMajorSelectActivity extends AppCompatActivity {
     private DrawerLayout drawerlayout;
     private Context context = this;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_user_main);
-
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        setContentView(R.layout.activity_user_major_select);
+        Toolbar toolbar = (Toolbar) findViewById(R.id.majorselecttoolbar);
         setSupportActionBar(toolbar);
         ActionBar actionBar = getSupportActionBar();
         actionBar.setHomeAsUpIndicator(R.drawable.ic_hamburger);
         actionBar.setDisplayHomeAsUpEnabled(true);
         actionBar.setDisplayShowTitleEnabled(false);
 
-        ConstraintLayout majorclub = (ConstraintLayout)findViewById(R.id.majorclubLayout);
-        ConstraintLayout mainclub = (ConstraintLayout)findViewById(R.id.mainclubLayout);
-        ConstraintLayout newclub = (ConstraintLayout)findViewById(R.id.newclubLayout);
-
-        ImageButton eventButton = (ImageButton)findViewById(R.id.usereventlist);
-
-        drawerlayout = (DrawerLayout) findViewById(R.id.drawer_layout_user_main);
-        NavigationView navigationView = (NavigationView) findViewById(R.id.nav_view_user_main);
+        drawerlayout = (DrawerLayout) findViewById(R.id.drawer_layout_user_major_select_xml);
+        NavigationView navigationView = (NavigationView) findViewById(R.id.nav_view_user_major_select);
         navigationView.setNavigationItemSelectedListener(new NavigationView.OnNavigationItemSelectedListener() {
             @Override
             public boolean onNavigationItemSelected(@NonNull MenuItem menuItem) {
@@ -85,41 +69,6 @@ public class UserMainActivity extends AppCompatActivity {
                 return true;
             }
         });
-
-        majorclub.setOnTouchListener(new View.OnTouchListener() {
-            @Override
-            public boolean onTouch(View v, MotionEvent event) {
-                Intent intent = new Intent(getApplicationContext(), UserMajorSelectActivity.class);
-                startActivity(intent);
-                return false;
-            }
-        });
-
-        mainclub.setOnTouchListener(new View.OnTouchListener() {
-            @Override
-            public boolean onTouch(View v, MotionEvent event) {
-                Intent intent = new Intent(getApplicationContext(), UserMainClubListActivity.class);
-                startActivity(intent);
-                return false;
-            }
-        });
-
-        newclub.setOnTouchListener(new View.OnTouchListener() {
-            @Override
-            public boolean onTouch(View v, MotionEvent event) {
-                Intent intent = new Intent(getApplicationContext(), UserNewClubListActivity.class);
-                startActivity(intent);
-                return false;
-            }
-        });
-
-        eventButton.setOnClickListener(new View.OnClickListener(){
-            public void onClick(View v){
-                Intent intent = new Intent(getApplicationContext(), UserNewClubListActivity.class);
-                startActivity(intent);
-            }
-        });
-
     }
 
     @Override
@@ -132,5 +81,4 @@ public class UserMainActivity extends AppCompatActivity {
         }
         return super.onOptionsItemSelected(item);
     }
-
 }
