+++ AjoudongFE/app/src/main/res/drawable-mdpi/major_select_button_shape.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_height="110dp" android:layout_width="110dp">
+
+    <stroke android:width="1dp" android:color="#707070" />
+
+</shape>
