+++ AjoudongFE/app/src/main/res/layout/activity_user_main_club_list.xml
@@ -1,9 +1,173 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.drawerlayout.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/drawer_layout_user_main_club_list"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-</androidx.constraintlayout.widget.ConstraintLayout>
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        tools:context=".UserMainClubListActivity">
+
+        <LinearLayout
+            android:id="@+id/gridLinear"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:orientation="vertical">
+
+            <androidx.appcompat.widget.Toolbar
+                android:id="@+id/mainclubtoolbar"
+                android:layout_width="match_parent"
+                android:layout_height="?attr/actionBarSize"
+                android:theme="@style/DarkToolBarTheme"/>
+
+            <HorizontalScrollView
+                android:layout_width="match_parent"
+                android:layout_height="50dp"
+                android:layout_marginBottom="5dp"
+                android:scrollbars="none"
+                android:background="@drawable/grid_scrollview_shape">
+
+                <LinearLayout
+                    android:layout_width="fill_parent"
+                    android:layout_height="fill_parent"
+                    android:orientation="horizontal">
+
+                    <Button
+                        android:id="@+id/cateAll"
+                        android:layout_width="50dp"
+                        android:layout_height="wrap_content"
+                        android:textColor="@color/grid_category_click"
+                        android:textColorHighlight="@color/grid_category_click"
+                        android:textColorLink="@color/grid_category_click"
+                        android:text="전체"
+                        />
+
+                    <Button
+                        android:id="@+id/cateSports"
+                        android:layout_width="90dp"
+                        android:layout_height="wrap_content"
+                        android:textColor="@color/grid_category_click"
+                        android:textColorHighlight="@color/grid_category_click"
+                        android:textColorLink="@color/grid_category_click"
+
+                        android:text="레저스포츠" />
+
+                    <Button
+                        android:id="@+id/cateReligion"
+                        android:layout_width="50dp"
+                        android:layout_height="wrap_content"
+                        android:textColor="@color/grid_category_click"
+                        android:textColorHighlight="@color/grid_category_click"
+                        android:textColorLink="@color/grid_category_click"
+
+                        android:text="종교" />
+
+                    <Button
+                        android:id="@+id/cateSocial"
+                        android:layout_width="75dp"
+                        android:layout_height="wrap_content"
+                        android:textColor="@color/grid_category_click"
+                        android:textColorHighlight="@color/grid_category_click"
+                        android:textColorLink="@color/grid_category_click"
+
+                        android:text="사회활동" />
+
+                    <Button
+                        android:id="@+id/cateCreate"
+                        android:layout_width="75dp"
+                        android:layout_height="wrap_content"
+                        android:textColor="@color/grid_category_click"
+                        android:textColorHighlight="@color/grid_category_click"
+                        android:textColorLink="@color/grid_category_click"
+
+                        android:text="창작전시" />
+
+                    <Button
+                        android:id="@+id/cateStudy"
+                        android:layout_width="50dp"
+                        android:layout_height="wrap_content"
+                        android:textColor="@color/grid_category_click"
+                        android:textColorHighlight="@color/grid_category_click"
+                        android:textColorLink="@color/grid_category_click"
+
+                        android:text="학습" />
+
+                    <Button
+                        android:id="@+id/catePE"
+                        android:layout_width="50dp"
+                        android:layout_height="wrap_content"
+                        android:textColor="@color/grid_category_click"
+                        android:textColorHighlight="@color/grid_category_click"
+                        android:textColorLink="@color/grid_category_click"
+
+                        android:text="체육" />
+
+                    <Button
+                        android:id="@+id/cateArt"
+                        android:layout_width="75dp"
+                        android:layout_height="wrap_content"
+                        android:textColor="@color/grid_category_click"
+                        android:textColorHighlight="@color/grid_category_click"
+                        android:textColorLink="@color/grid_category_click"
+
+                        android:text="언행예술" />
+
+                    <Button
+                        android:id="@+id/cateSub"
+                        android:layout_width="75dp"
+                        android:layout_height="wrap_content"
+                        android:textColor="@color/grid_category_click"
+                        android:textColorHighlight="@color/grid_category_click"
+                        android:textColorLink="@color/grid_category_click"
+
+                        android:text="준동아리" />
+                </LinearLayout>
+            </HorizontalScrollView>
+
+            <androidx.constraintlayout.widget.ConstraintLayout
+                android:layout_width="match_parent"
+                android:layout_height="147dp"
+                android:layout_weight="0.2">
+
+                <TextView
+                    android:id="@+id/textView4"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_marginEnd="10dp"
+                    android:text="↑↓정렬"
+                    app:layout_constraintBottom_toBottomOf="parent"
+                    app:layout_constraintEnd_toEndOf="parent" />
+            </androidx.constraintlayout.widget.ConstraintLayout>
+
+            <GridView
+                android:id="@+id/gridView01"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:layout_weight="0.2"
+                android:verticalSpacing="10dp"
+                android:horizontalSpacing="10dp"
+                android:numColumns="3"
+                android:paddingLeft="10dp"
+                android:paddingRight="10dp"
+                android:paddingTop="10dp"
+                android:paddingBottom="10dp"
+
+
+                />
+
+
+        </LinearLayout>
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+    <com.google.android.material.navigation.NavigationView
+        android:id="@+id/nav_view_user_main_club_list"
+        android:layout_width="wrap_content"
+        android:layout_height="match_parent"
+        android:layout_gravity="start"
+        android:fitsSystemWindows="true"
+        app:headerLayout="@layout/navi_user_header"
+        app:menu="@menu/navi_user_menu" />
+</androidx.drawerlayout.widget.DrawerLayout>
